:recycle: refactor: rename util package to utils
