+++ app/services/generate_service.py
@@ -1,7 +1,7 @@
 import ollama
 
 from app.schemas.generate_dto import GenerateDTO
-from app.util.json_extractor import extract_json_from_response
+from app.utils.json_extractor import extract_json_from_response
 
 
 class GenerateService:
